+++ passport/app/Account.php
@@ -1,0 +1,46 @@
+<?php
+
+namespace App;
+
+use Illuminate\Notifications\Notifiable;
+use Illuminate\Foundation\Auth\User as Authenticatable;
+use Laravel\Passport\HasApiTokens;
+
+class Account extends Authenticatable
+{
+    use Notifiable, HasApiTokens;
+
+    protected $table = 'Account';
+
+    /**
+     * The attributes that are mass assignable.
+     *
+     * @var array
+     */
+    protected $fillable = [
+        'user_name', 'email', 'password', 'role_id', 'last_ip', 'login_attempts_count', 'locked_to'
+    ];
+
+    /**
+     * The attributes that should be hidden for arrays.
+     *
+     * @var array
+     */
+    protected $hidden = [
+        'password', 'remember_token',
+    ];
+
+    /**
+     * The attributes that should be cast to native types.
+     *
+     * @var array
+     */
+    protected $casts = [
+        'email_verified_at' => 'datetime',
+    ];
+    
+    public function role()
+    {
+        return $this->hasOne(Role::class);
+    }
+}
