+++ passport/app/Post.php
@@ -11,4 +11,9 @@ class Post extends Model
     protected $fillable = [
         'title', 'path', 'content', 'description'
     ];
+
+    public function user()
+    {
+        return $this->belongsTo(User::class);
+    }
 }
