+++ passport/app/Title.php
@@ -1,0 +1,17 @@
+<?php
+
+namespace App;
+
+use Illuminate\Database\Eloquent\Model;
+
+class Title extends Model
+{
+    protected $fillable = [
+        'name'
+    ];
+
+    public function users()
+    {
+        return $this->hasMany(User::class);
+    }
+}
