+++ passport/app/User.php
@@ -2,50 +2,26 @@
 
 namespace App;
 
-use Illuminate\Notifications\Notifiable;
-use Illuminate\Foundation\Auth\User as Authenticatable;
-use Laravel\Passport\HasApiTokens;
+use Illuminate\Database\Eloquent\Model;
+use App\Account;
 
-class User extends Authenticatable
+class User extends Model
 {
-    use Notifiable, HasApiTokens;
 
-    protected $table = 'Account';
+    protected $table = 'User';
 
-    /**
-     * The attributes that are mass assignable.
-     *
-     * @var array
-     */
     protected $fillable = [
-        'user_name', 'email', 'password', 'role'
+        'account_id', 'title_id', 'avatar', 'display_name', 'experience', 'rows_scrolled', 'custom_path', 'created_at', 'updated_at', 'deleted_at'
     ];
 
-    /**
-     * The attributes that should be hidden for arrays.
-     *
-     * @var array
-     */
-    protected $hidden = [
-        'password', 'remember_token',
-    ];
-
-    /**
-     * The attributes that should be cast to native types.
-     *
-     * @var array
-     */
-    protected $casts = [
-        'email_verified_at' => 'datetime',
-    ];
-
-    public function posts()
+    public function title()
     {
-        return $this->hasMany(Post::class);
+        return $this->belongsTo(Title::class);
     }
 
-    public function role()
+    public function account()
     {
-        return $this->hasOne(Role::class);
+        return $this->belongsTo(Account::class);
     }
+
 }
